requirements.txt
Django>=3.1,<3.2
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.5,<2.9.0
Pillow>=5.3.0,<5.4.0

flake8>=3.8.3,<3.9.0 